Removed explicit caching of decorated service types in favour of allowing the caching of ServiceRegistrationInfo for decorators instead.
