+++ src/Autofac/Features/OpenGenerics/OpenGenericRegistrationSource.cs
@@ -41,8 +41,6 @@ namespace Autofac.Features.OpenGenerics
         private readonly RegistrationData _registrationData;
         private readonly ReflectionActivatorData _activatorData;
 
-        internal IEnumerable<Service> Services => _registrationData.Services;
-
         public OpenGenericRegistrationSource(
             RegistrationData registrationData,
             ReflectionActivatorData activatorData)
